Combine conda_env.exceptions with conda.exceptions
