+++ conda/cli/main_compare.py
@@ -11,16 +11,16 @@ from .common import stdout_json
 from ..base.context import context
 from ..common.compat import text_type
 from ..core.prefix_data import PrefixData
+from ..exceptions import EnvironmentLocationNotFound, SpecNotFound
 from ..gateways.connection.session import CONDA_SESSION_SCHEMES
 from ..gateways.disk.test import is_conda_environment
-from conda_env import exceptions, specs
 from ..models.match_spec import MatchSpec
+from conda_env import specs
 
 log = logging.getLogger(__name__)
 
 def get_packages(prefix):
     if not os.path.isdir(prefix):
-        from ..exceptions import EnvironmentLocationNotFound
         raise EnvironmentLocationNotFound(prefix)
 
     return sorted(PrefixData(prefix, pip_interop_enabled=True).iter_records(),
@@ -58,7 +58,6 @@ with matching version and build string.")
 def execute(args, parser):
     prefix = context.target_prefix
     if not is_conda_environment(prefix):
-        from ..exceptions import EnvironmentLocationNotFound
         raise EnvironmentLocationNotFound(prefix)
 
     try:
@@ -73,7 +72,7 @@ def execute(args, parser):
 
         if args.prefix is None and args.name is None:
             args.name = env.name
-    except exceptions.SpecNotFound:
+    except SpecNotFound:
         raise
 
     active_pkgs = dict(map(_get_name_tuple, get_packages(prefix)))
