+++ conda/exceptions.py
@@ -954,6 +954,39 @@ class NoSpaceLeftError(CondaError):
         super(NoSpaceLeftError, self).__init__(message, caused_by=caused_by, **kwargs)
 
 
+class CondaEnvException(CondaError):
+    def __init__(self, message, *args, **kwargs):
+        msg = "%s" % message
+        super(CondaEnvException, self).__init__(msg, *args, **kwargs)
+
+
+class EnvironmentFileNotFound(CondaEnvException):
+    def __init__(self, filename, *args, **kwargs):
+        msg = "'{}' file not found".format(filename)
+        self.filename = filename
+        super(EnvironmentFileNotFound, self).__init__(msg, *args, **kwargs)
+
+
+class EnvironmentFileExtensionNotValid(CondaEnvException):
+    def __init__(self, filename, *args, **kwargs):
+        msg = "'{}' file extension must be one of '.txt', '.yaml' or '.yml'".format(filename)
+        self.filename = filename
+        super(EnvironmentFileExtensionNotValid, self).__init__(msg, *args, **kwargs)
+
+
+class EnvironmentFileNotDownloaded(CondaError):
+    def __init__(self, username, packagename, *args, **kwargs):
+        msg = "{}/{} file not downloaded".format(username, packagename)
+        self.username = username
+        self.packagename = packagename
+        super(EnvironmentFileNotDownloaded, self).__init__(msg, *args, **kwargs)
+
+
+class SpecNotFound(CondaError):
+    def __init__(self, msg, *args, **kwargs):
+        super(SpecNotFound, self).__init__(msg, *args, **kwargs)
+
+
 def maybe_raise(error, context):
     if isinstance(error, CondaMultiError):
         groups = groupby(lambda e: isinstance(e, ClobberError), error.errors)
