+++ conda_env/specs/yaml_file.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
+from conda.exceptions import EnvironmentFileNotFound
+
 from .. import env
-from ..exceptions import EnvironmentFileNotFound
 
 
 class YamlFileSpec(object):
