+++ tests/conda_env/test_cli.py
@@ -18,11 +18,15 @@ from conda.common.compat import odict
 from conda.common.io import captured
 from conda.common.serialize import yaml_safe_load
 from conda.core.envs_manager import list_all_known_prefixes
-from conda.exceptions import EnvironmentLocationNotFound
+from conda.exceptions import (
+    EnvironmentLocationNotFound,
+    CondaEnvException,
+    EnvironmentFileExtensionNotValid,
+    EnvironmentFileNotFound,
+)
 from conda.gateways.disk.delete import rm_rf
 from conda.utils import massage_arguments
 from conda_env.cli.main import create_parser, do_call as do_call_conda_env
-from conda_env.exceptions import CondaEnvException, EnvironmentFileExtensionNotValid, EnvironmentFileNotFound
 
 environment_1 = '''
 name: env-1
